renamed pate_cmd_line to align
