+++ src/paste3/align.py
